renamed modules and redistributed utilities between obsdata and utils
